+++ src/routes/scholar/[id]/+page.ts
@@ -1,5 +1,6 @@
-/** @type {import('./$types').PageLoad} */
-export async function load({ parent, params }) {
+import type { PageLoad } from './$types';
+
+export const load: PageLoad = async ({ parent, params }) => {
 	const { supabase } = await parent();
 
 	const { data, error } = await supabase.from('scholars').select().eq('id', params.id).single();
@@ -7,4 +8,4 @@ export async function load({ parent, params }) {
 	return {
 		scholar: data && error === null ? data : null
 	};
-}
+};